9 try

+++ EX/ex02_math/maths.py
@@ -25,4 +25,4 @@ def add_fractions(a: int, b: int, c: int, d: int) -> str:
     """Add_fractions function."""
     x = a / b
     y = c / d
-    return str(int(x/y))
+    return str(x/y)
